% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\alias{psps_import}
\title{Import Functions for the PSPS data}
\usage{
psps_import(path = NULL, years = 2010:2019, ...)
}
\arguments{
\item{path}{path to the directory where the downloaded data is stored.}

\item{years}{which years to import}

\item{...}{additional arguments passed to \code{data.table::fread}}
}
\value{
A \code{\link[data.table]{data.table}}
}
\description{
Import Functions for the PSPS data
}
\examples{
\dontrun{

library(cms.psps)

# Check that the data is as expected
check <- psps_md5sum()
if (!all(check$check)) {
  psps_download()
  psps_unzip()
}

psps_2010 <- psps_import(year = 2010)
psps_2010

psps_2018_2019 <- psps_import(years = 2018:2019)
psps_2018_2019

}

}
