% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{psps_unzip}
\alias{psps_unzip}
\title{Unzip Downloaded Data Sets}
\usage{
psps_unzip(path = NULL)
}
\arguments{
\item{path}{directory to the downloaded PSPS files.}

\item{...}{not currently used.}
}
\description{
Unzip Downloaded Data Sets
}
