% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\alias{psps_import_2010}
\title{Import Functions for the PSPS data}
\usage{
psps_import_2010(path = NULL, ...)
}
\arguments{
\item{path}{path to the directory where the downloaded data is stored.}

\item{...}{additional arguments passed to \code{data.table::fread}}
}
\value{
A \code{\link[data.table]{data.table}}
}
\description{
Import Functions for the PSPS data
}
