% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{psps_download}
\alias{psps_download}
\title{Download the PSPS Data}
\usage{
psps_download(path = NULL, years = 2010:2019, ...)
}
\arguments{
\item{path}{directory to save the downloaded files too.}

\item{years}{years to download}

\item{...}{not currently used.}
}
\description{
Download the public files provided by CMS.
}
