% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{psps_md5sum}
\alias{psps_md5sum}
\title{Verify PSPS .zip files via MD5SUM}
\usage{
psps_md5sum(path = NULL, years = 2010:2019, ...)
}
\arguments{
\item{path}{directory to the downloaded PSPS files.}

\item{years}{data sets to verify}

\item{...}{not currently used.}
}
\description{
Verify PSPS .zip files via MD5SUM
}
