% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_import.R
\name{spark_import}
\alias{spark_import}
\alias{spark_psps_import_2010}
\alias{spark_psps_import_2011}
\alias{spark_psps_import_2012}
\alias{spark_psps_import_2013}
\alias{spark_psps_import_2014}
\alias{spark_psps_import_2015}
\alias{spark_psps_import_2016}
\alias{spark_psps_import_2017}
\alias{spark_psps_import_2018}
\title{Spark Import Functions for the PSPS data}
\usage{
spark_psps_import_2010(sc, path = NULL, envir = .GlobalEnv, ...)

spark_psps_import_2011(sc, path = NULL, envir = .GlobalEnv, ...)

spark_psps_import_2012(sc, path = NULL, envir = .GlobalEnv, ...)

spark_psps_import_2013(sc, path = NULL, envir = .GlobalEnv, ...)

spark_psps_import_2014(sc, path = NULL, envir = .GlobalEnv, ...)

spark_psps_import_2015(sc, path = NULL, envir = .GlobalEnv, ...)

spark_psps_import_2016(sc, path = NULL, envir = .GlobalEnv, ...)

spark_psps_import_2017(sc, path = NULL, envir = .GlobalEnv, ...)

spark_psps_import_2018(sc, path = NULL, envir = .GlobalEnv, ...)
}
\arguments{
\item{sc}{a sparklyr cluster}

\item{path}{path to the directory where the downloaded data is stored.}

\item{envir}{the envir to assign the \code{tbl_spark} object too, defatuls to
the global environment.}

\item{...}{arguments passed to \code{\link[sparklyr]{spark_read_csv}} (for
the import of 2017 data) or to \code{\link[sparklyr]{spark_read_text}} for
the years 2004 through 2016.}
}
\value{
A \code{tbl_spark} object of the name psps_<year> will be in the
\code{envir}.
}
\description{
Importing the PSPS data into spark via sparklyr
}
\examples{
\donttest{

library(cms.psps)

# Check that the data is as expected
check <- psps_md5sum()
if (!all(check$check)) {
  psps_download()
  psps_unzip()
}

# Set up spark if needed
# sparklyr::spark_install()

sc <- sparklyr::spark_connect("local")

ls()

spark_psps_import_2010(sc, memory = FALSE)
psps_2010

spark_psps_import_2011(sc, memory = FALSE)
psps_2011

spark_psps_import_2012(sc, memory = FALSE)
psps_2012

spark_psps_import_2013(sc, memory = FALSE)
psps_2013

spark_psps_import_2014(sc, memory = FALSE)
psps_2014

spark_psps_import_2015(sc, memory = FALSE)
psps_2015

# Be aware: There are some non utf8 characters in the data.  Be aware and
# explore and test your data exports.  In the example below there is a
library(magrittr)
psps_2015 \%>\%
  dplyr::group_by(HCPCS_INITIAL_MODIFIER_CD) \%>\%
  dplyr::summarize(n()) \%>\%
  dplyr::arrange(HCPCS_INITIAL_MODIFIER_CD) \%>\%
  dplyr::collect() \%>\%
  dplyr::filter(grepl("^G.*", HCPCS_INITIAL_MODIFIER_CD)) \%>\%
  print(n = Inf)

spark_psps_import_2016(sc, memory = FALSE)
psps_2016

spark_psps_import_2017(sc, memory = FALSE)
psps_2017

spark_psps_import_2018(sc, memory = FALSE)
psps_2018

ls()

sparklyr::spark_disconnect(sc)

}


}
